+++ src/container/mod.rs
@@ -5,4 +5,3 @@ pub mod uidmap;
 pub mod mount;
 pub mod root;
 pub mod nsutil;
-pub mod vagga;