+++ src/launcher/build.rs
@@ -1,17 +1,73 @@
+use std::env;
 use std::io::{stdout, stderr};
 
 use argparse::{ArgumentParser, Store, StoreTrue};
+use unshare::{Command, Namespace};
 
 use config::Config;
+use process_util::{capture_fd3, set_uidmap};
+use container::uidmap::get_max_uidmap;
 
-use super::user;
 
-
-pub fn build_container(config: &Config, name: &String) -> Result<(), String> {
-    build_command(config, vec!(name.clone())).map(|_| ())
+pub fn build_container(_config: &Config, name: &String)
+    -> Result<String, String>
+{
+    build_internal(name, &[])
 }
 
-pub fn build_command(_config: &Config, args: Vec<String>)
+/// Similar to build_container but never actually builds
+pub fn get_version(name: &str) -> Result<String, String> {
+    let mut cmd = Command::new("/proc/self/exe");
+    cmd.arg0("vagga_wrapper");
+    cmd.arg("_version_hash");
+    cmd.arg("--short");
+    cmd.arg("--fd3");
+    cmd.arg(name);
+    cmd.env_clear();
+    if let Ok(x) = env::var("RUST_LOG") {
+        cmd.env("RUST_LOG", x);
+    }
+    if let Ok(x) = env::var("RUST_BACKTRACE") {
+        cmd.env("RUST_BACKTRACE", x);
+    }
+    if let Ok(x) = env::var("HOME") {
+        cmd.env("VAGGA_USER_HOME", x);
+    }
+    cmd.unshare(
+        [Namespace::Mount, Namespace::Ipc, Namespace::Pid].iter().cloned());
+    set_uidmap(&mut cmd, &get_max_uidmap().unwrap(), true);
+
+    capture_fd3(cmd)
+    .and_then(|x| String::from_utf8(x)
+                  .map_err(|e| format!("Can't decode version: {}", e)))
+}
+
+fn build_internal(name: &str, args: &[String]) -> Result<String, String> {
+    let mut cmd = Command::new("/proc/self/exe");
+    cmd.arg0("vagga_wrapper");
+    cmd.arg("_build");
+    cmd.arg(name);
+    cmd.args(&args);
+    cmd.env_clear();
+    if let Ok(x) = env::var("RUST_LOG") {
+        cmd.env("RUST_LOG", x);
+    }
+    if let Ok(x) = env::var("RUST_BACKTRACE") {
+        cmd.env("RUST_BACKTRACE", x);
+    }
+    if let Ok(x) = env::var("HOME") {
+        cmd.env("VAGGA_USER_HOME", x);
+    }
+    cmd.unshare(
+        [Namespace::Mount, Namespace::Ipc, Namespace::Pid].iter().cloned());
+    set_uidmap(&mut cmd, &get_max_uidmap().unwrap(), true);
+
+    capture_fd3(cmd)
+    .and_then(|x| String::from_utf8(x)
+                  .map_err(|e| format!("Can't decode version: {}", e)))
+}
+
+pub fn build_command(_config: &Config, mut args: Vec<String>)
     -> Result<i32, String>
 {
     let mut name: String = "".to_string();
@@ -37,9 +93,9 @@ pub fn build_command(_config: &Config, args: Vec<String>)
             }
         }
     }
-    match user::run_wrapper(None, "_build".to_string(), args, true) {
-        Ok(0) => Ok(0),
-        Ok(x) => Err(format!("Build returned {}", x)),
-        Err(e) => Err(e),
-    }
+    assert!(args.remove(0) == name);
+
+    build_internal(&name, &args)
+    .map(|v| debug!("Container {:?} build with version {:?}", name, v))
+    .map(|()| 0)
 }