+++ src/launcher/underscore.rs
@@ -2,7 +2,7 @@ use std::io::{stdout, stderr};
 use std::path::Path;
 
 use argparse::{ArgumentParser};
-use argparse::{StoreTrue, List, StoreOption, Store};
+use argparse::{StoreTrue, List, StoreOption, Store, StoreFalse};
 use unshare::Namespace;
 
 use config::Config;
@@ -10,11 +10,11 @@ use container::nsutil::{set_namespace};
 
 use super::user;
 use super::network;
-use super::build::build_container;
+use super::build::{build_container, get_version};
 
 
 pub fn run_command(config: &Config, workdir: &Path, cmdname: String,
-    mut args: Vec<String>)
+    mut args: Vec<String>, mut allow_build: bool)
     -> Result<i32, String>
 {
     let mut cmdargs = Vec::<String>::new();
@@ -32,6 +32,10 @@ pub fn run_command(config: &Config, workdir: &Path, cmdname: String,
                  Currently we use hard-linked copy of the container, so it's
                  dangerous for some operations. Still it's ok for installing
                  packages or similar tasks");
+        ap.refer(&mut allow_build)
+            .add_option(&["--no-build"], StoreFalse, "
+                Do not build container even if it is out of date. Return error
+                code 29 if it's out of date.");
         ap.refer(&mut container)
             .add_argument("container", Store,
                 "Container to run command in")
@@ -51,12 +55,17 @@ pub fn run_command(config: &Config, workdir: &Path, cmdname: String,
         }
     }
     args.remove(0);
-    try!(build_container(config, &container));
-    let res = user::run_wrapper(Some(workdir), cmdname, args, true);
+    let ver = if allow_build {
+        try!(build_container(config, &container))
+    } else {
+        format!("{}.{}", &container, try!(get_version(&container)))
+    };
+    let res = user::run_wrapper(Some(workdir), cmdname, args,
+        true, Some(&ver));
 
     if copy {
         match user::run_wrapper(Some(workdir), "_clean".to_string(),
-            vec!("--transient".to_string()), true)
+            vec!("--transient".to_string()), true, None)
         {
             Ok(0) => {}
             x => warn!(
@@ -68,7 +77,7 @@ pub fn run_command(config: &Config, workdir: &Path, cmdname: String,
 }
 
 pub fn run_in_netns(config: &Config, workdir: &Path, cname: String,
-    mut args: Vec<String>)
+    mut args: Vec<String>, mut allow_build: bool)
     -> Result<i32, String>
 {
     let mut cmdargs = vec!();
@@ -84,6 +93,10 @@ pub fn run_in_netns(config: &Config, workdir: &Path, cname: String,
                 Run in the namespace of the process with PID.
                 By default you get shell in the \"gateway\" namespace.
                 ");
+        ap.refer(&mut allow_build)
+            .add_option(&["--no-build"], StoreFalse, "
+                Do not build container even if it is out of date. Return error
+                code 29 if it's out of date.");
         ap.refer(&mut container)
             .add_argument("container", Store,
                 "Container to run command in")
@@ -103,11 +116,15 @@ pub fn run_in_netns(config: &Config, workdir: &Path, cname: String,
         }
     }
     cmdargs.insert(0, container.clone());
-    try!(build_container(config, &container));
+    let ver = if allow_build {
+        try!(build_container(config, &container))
+    } else {
+        format!("{}.{}", &container, try!(get_version(&container)))
+    };
     try!(network::join_gateway_namespaces());
     if let Some::<i32>(pid) = pid {
         try!(set_namespace(format!("/proc/{}/ns/net", pid), Namespace::Net)
             .map_err(|e| format!("Error setting networkns: {}", e)));
     }
-    user::run_wrapper(Some(workdir), cname, cmdargs, false)
+    user::run_wrapper(Some(workdir), cname, cmdargs, false, Some(&ver))
 }