+++ src/launcher/user.rs
@@ -8,22 +8,23 @@ use config::command::MainCommand;
 use config::command::{CommandInfo, Networking, WriteMode};
 
 use super::supervisor;
-use super::build::build_container;
+use super::build::{build_container, get_version};
 use process_util::{convert_status, set_uidmap};
 use container::uidmap::get_max_uidmap;
 
 
 pub fn run_user_command(config: &Config, workdir: &Path,
-    cmd: String, args: Vec<String>)
+    cmd: String, args: Vec<String>, allow_build: bool)
     -> Result<i32, String>
 {
     match config.commands.get(&cmd) {
         None => Err(format!("Command {} not found. \
                     Run vagga without arguments to see the list.", cmd)),
         Some(&MainCommand::Command(ref info))
-        => run_simple_command(config, info, workdir, cmd, args),
+        => run_simple_command(config, info, workdir, cmd, args, allow_build),
         Some(&MainCommand::Supervise(ref sup))
-        => supervisor::run_supervise_command(config, workdir, sup, cmd, args),
+        => supervisor::run_supervise_command(config, workdir, sup,
+            cmd, args, allow_build),
     }
 }
 
@@ -52,19 +53,24 @@ pub fn common_child_command_env(cmd: &mut Command, workdir: Option<&Path>) {
 }
 
 pub fn run_simple_command(config: &Config, cfg: &CommandInfo,
-    workdir: &Path, cmdname: String, args: Vec<String>)
+    workdir: &Path, cmdname: String, args: Vec<String>, allow_build: bool)
     -> Result<i32, String>
 {
     if let Some(_) = cfg.network {
         return Err(format!(
             "Network is not supported for !Command use !Supervise"))
     }
-    try!(build_container(config, &cfg.container));
-    let res = run_wrapper(Some(workdir), cmdname, args, cfg.network.is_none());
+    let ver = if allow_build {
+        try!(build_container(config, &cfg.container))
+    } else {
+        format!("{}.{}", &cfg.container, try!(get_version(&cfg.container)))
+    };
+    let res = run_wrapper(Some(workdir), cmdname, args,
+        cfg.network.is_none(), Some(&ver));
 
     if cfg.write_mode != WriteMode::read_only {
         match run_wrapper(Some(workdir), "_clean".to_string(),
-            vec!("--transient".to_string()), true)
+            vec!("--transient".to_string()), true, None)
         {
             Ok(0) => {}
             x => warn!(
@@ -77,11 +83,15 @@ pub fn run_simple_command(config: &Config, cfg: &CommandInfo,
 
 // TODO(tailhook) run not only for simple commands
 pub fn run_wrapper(workdir: Option<&Path>, cmdname: String, args: Vec<String>,
-    userns: bool)
+    userns: bool, root: Option<&str>)
     -> Result<i32, String>
 {
     let mut cmd = Command::new("/proc/self/exe");
     cmd.arg0("vagga_wrapper");
+    if let Some(root) = root {
+        cmd.arg("--root");
+        cmd.arg(root);
+    };
     cmd.arg(&cmdname);
     cmd.args(&args);
     cmd.env_clear();