+++ src/wrapper/supervise.rs
@@ -12,7 +12,6 @@ use super::super::config::command::{SuperviseInfo, CommandInfo, WriteMode};
 use super::super::config::command::ChildCommand as CC;
 use super::super::container::uidmap::{map_users};
 use super::super::container::uidmap::Uidmap::Ranges;
-use super::super::container::vagga::container_ver;
 use super::Wrapper;
 use super::util::find_cmd;
 use super::setup;
@@ -99,7 +98,7 @@ fn supervise_child_command(cmdname: &String, name: &String, bridge: bool,
         WriteMode::transient_hard_link_copy
         => setup::WriteMode::TransientHardlinkCopy(pid),
     };
-    let cont_ver = try!(container_ver(&command.container));
+    let cont_ver = wrapper.root.as_ref().unwrap();
     try!(setup::setup_filesystem(cconfig, write_mode, &cont_ver));
 
     try!(_write_hosts(supervise));